requirements.txt
python-dotenv==1.1.1
pydantic==2.11.7
openai==1.97.0
langchain==0.3.26
langchain-community==0.3.27
gradio==5.38.0

mcp==1.12.2
langgraph==0.5.4
langchain-openai==0.3.28